introduce moreFunctions tab
