+++ frontend/static/frontend/main.js.LICENSE.txt
@@ -54,10 +54,6 @@
   !*** ./src/components/Cookies.js ***!
   \***********************************/
 
-/*!***********************************!*\
-  !*** ./src/components/Map/Map.js ***!
-  \***********************************/
-
 /*!************************************!*\
   !*** ./src/actions/authActions.js ***!
   \************************************/
@@ -95,6 +91,10 @@
   \*************************************/
 
 /*!*************************************!*\
+  !*** ./src/components/Map/index.js ***!
+  \*************************************/
+
+/*!*************************************!*\
   !*** ./src/reducers/authReducer.js ***!
   \*************************************/
 
@@ -119,7 +119,7 @@
   \***************************************/
 
 /*!***************************************!*\
-  !*** ./src/components/About/About.js ***!
+  !*** ./src/components/About/index.js ***!
   \***************************************/
 
 /*!***************************************!*\
@@ -219,6 +219,10 @@
   \*****************************************/
 
 /*!*****************************************!*\
+  !*** ./src/components/Auth/Register.js ***!
+  \*****************************************/
+
+/*!*****************************************!*\
   !*** ./src/components/Navbar/Navbar.js ***!
   \*****************************************/
 
@@ -300,6 +304,10 @@
 
 /*!*******************************************!*\
   !*** ./node_modules/stylis/src/Parser.js ***!
+  \*******************************************/
+
+/*!*******************************************!*\
+  !*** ./src/components/Functions/index.js ***!
   \*******************************************/
 
 /*!*******************************************!*\
@@ -531,6 +539,10 @@
   \*************************************************/
 
 /*!*************************************************!*\
+  !*** ./src/components/Private/Profile/index.js ***!
+  \*************************************************/
+
+/*!*************************************************!*\
   !*** ./src/i18n/messages/de-De/translations.js ***!
   \*************************************************/
 
@@ -583,10 +595,6 @@
   \**************************************************/
 
 /*!**************************************************!*\
-  !*** ./src/components/Auth/Register/Register.js ***!
-  \**************************************************/
-
-/*!**************************************************!*\
   !*** ./src/components/Navbar/ProfileDropdown.js ***!
   \**************************************************/
 
@@ -632,10 +640,6 @@
 
 /*!***************************************************!*\
   !*** ./node_modules/zxcvbn/lib/time_estimates.js ***!
-  \***************************************************/
-
-/*!***************************************************!*\
-  !*** ./src/components/Private/Profile/Profile.js ***!
   \***************************************************/
 
 /*!****************************************************!*\
@@ -748,10 +752,6 @@
 
 /*!*****************************************************!*\
   !*** ./node_modules/zxcvbn/lib/adjacency_graphs.js ***!
-  \*****************************************************/
-
-/*!*****************************************************!*\
-  !*** ./src/components/Auth/Register/StepWrapper.js ***!
   \*****************************************************/
 
 /*!******************************************************!*\
