+++ users/serializers.py
@@ -37,10 +37,13 @@ class UserRegistrationSerializer(serializers.ModelSerializer):
                                      validate_password], style={'input_type': 'password', 'placeholder': 'Password'})
     password2 = serializers.CharField(required=True, style={
                                       'input_type': 'password', 'placeholder': 'Confirm password'})
+    firstname = serializers.CharField()
+    lastname = serializers.CharField()
 
     class Meta:
         model = User
-        fields = ('username',  'email', 'password', 'password2')
+        fields = ('username',  'email', 'password',
+                  'password2', 'first_name', 'last_name')
 
     def validate(self, attrs):
         if attrs['password'] != attrs['password2']:
@@ -52,7 +55,9 @@ class UserRegistrationSerializer(serializers.ModelSerializer):
     def create(self, validated_data):
         user = User.objects.create(
             username=validated_data['username'],
-            email=validated_data['email']
+            email=validated_data['email'],
+            first_name=validated_data['firstname'],
+            last_name=validated_data['lastname']
         )
 
         user.set_password(validated_data['password'])
